Fixed for 1.1.6

+++ install/index.php
@@ -3,7 +3,7 @@ header("Cache-Control: no-cache, no-store, must-revalidate"); /* HTTP 1.1. */
 header("Pragma: no-cache"); /* HTTP 1.0. */
 header("Expires: 0"); /* Proxies. */
 if (file_exists('../config.inc.php')) {
-    header("Location: ../?nocache=".time());
+    header("Location: ../?nocache=" . time());
     exit();
 }
 /**
@@ -28,7 +28,7 @@ if (file_exists('../config.inc.php')) {
 require './include/Database.php';
 $success = 0;
 $chk_pass = 0;
-if (!empty($_POST) && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] && $_POST['dbpass'] && $_POST['dbname']) {
+if (!empty($_POST) && $_POST['submitbtn'] && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] && $_POST['dbpass'] && $_POST['dbname']) {
     /* Prepare Input Data */
     /* $dbdsn = $_POST['dbdsn']; */
     $url_replace = array('https://', 'http://');
@@ -50,7 +50,7 @@ if (!empty($_POST) && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] 
         $filename = 'cszcms_app.sql';
         $db->mysqli_multi_query_file($mysqli, $filename);
 
-        $insert_user = "INSERT INTO `user_admin` (`user_admin_id`, `name`, `email`, `password`, `user_type`, `active`, `md5_hash`, `md5_lasttime`, `timestamp_create`, `timestamp_update`) VALUES (1, 'Admin User', '" . $email . "', '" . sha1(md5($_POST['password'])) . "', 'admin', 1, '" . md5(time() + mt_rand(1, 99999999)) . "', NOW(), NOW(), NOW())";
+        $insert_user = "INSERT INTO `user_admin` (`user_admin_id`, `name`, `email`, `password`, `user_type`, `active`, `md5_hash`, `md5_lasttime`, `timestamp_create`, `timestamp_update`) VALUES (1, 'Admin User', '" . $email . "', '" . hash('sha512', sha1(md5($_POST['password']))) . "', 'admin', 1, '" . md5(time() + mt_rand(1, 99999999)) . "', NOW(), NOW(), NOW())";
         $mysqli->query($insert_user);
         $update_sql = "UPDATE `settings` SET `default_email` = '" . $email . "' WHERE `settings_id` = 1";
         $mysqli->query($update_sql);
@@ -70,8 +70,6 @@ if (!empty($_POST) && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] 
         $config_file = '../config.inc.php';
         $config_txt = "<?php \n";
         $config_txt .= "defined('FCPATH') OR exit('No direct script access allowed'); \n\n";
-        $config_txt .= "/* Database DSN */ \n";
-        $config_txt .= "define('DB_DSN', ''); \n\n";
         $config_txt .= "/* Database Host */ \n";
         $config_txt .= "define('DB_HOST', '" . $dbhost . "'); \n\n";
         $config_txt .= "/* Database Username */ \n";
@@ -80,8 +78,6 @@ if (!empty($_POST) && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] 
         $config_txt .= "define('DB_PASS', '" . $dbpass . "'); \n\n";
         $config_txt .= "/* Database Name */ \n";
         $config_txt .= "define('DB_NAME', '" . $dbname . "'); \n\n";
-        $config_txt .= "/* Database Driver */ \n";
-        $config_txt .= "define('DB_DRIVER', 'mysqli'); \n\n";
         $config_txt .= "/* Base URL */ \n";
         $config_txt .= "define('BASE_URL', '" . $baseurl . "'); \n\n";
         $config_txt .= "/* Email Domain */ \n";
@@ -111,7 +107,7 @@ if (!empty($_POST) && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] 
         <meta http-equiv="X-UA-Compatible" content="IE=edge" />
         <meta http-equiv="Content-type" content="text/html; charset=utf-8" />
         <link href="../templates/admin/imgs/favicon.ico" rel="shortcut icon" type="image/ico" />
-        <title>Install System | CSZ-CMS System</title>
+        <title>Install System | CSZ CMS System</title>
 
         <!-- Bootstrap Core CSS -->
         <link href="../assets/css/bootstrap.min.css" rel="stylesheet" type="text/css" />
@@ -171,7 +167,7 @@ if (!empty($_POST) && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] 
                         <div class="text-center"><a href="https://www.cszcms.com" target="_blank" title="CSZ CMS Official Website"><img alt="Site Logo" class="site-logo" src="assets/images/logo.png"></a></div>
                         <br><br>
                         <div class="text-center">
-                            <div class="well"><h3 class="form-signin-heading">Install CSZ-CMS</h3></div>
+                            <div class="well"><h3 class="form-signin-heading">Install CSZ CMS</h3></div>
                         </div>
                     </div>
                 </div>
@@ -179,62 +175,78 @@ if (!empty($_POST) && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] 
                     <div class="col-md-12">
                         <b><u>System Checking</u></b><br><br>
                         <b>
-                                PHP 5.3.7 or higher [<?php if (version_compare(phpversion(), '5.3.7', '>=') !== FALSE) {
+                            PHP 5.3.7 or higher [<?php
+                            if (version_compare(phpversion(), '5.3.7', '>=') !== FALSE) {
                                 echo '<span class="success">PASS</span>';
                                 $php = 1;
                             } else {
                                 echo '<span class="error">FAIL</span>';
                                 $php = 0;
-                            } ?>]<br>
-                                MySQLi Driver [<?php if (extension_loaded('mysqli') !== FALSE) {
+                            }
+                            ?>]<br>
+                            MySQLi Driver [<?php
+                            if (extension_loaded('mysqli') !== FALSE) {
                                 echo '<span class="success">PASS</span>';
                                 $sqli = 1;
                             } else {
                                 echo '<span class="error">FAIL</span>';
                                 $sqli = 0;
-                            } ?>]<br>
-                                cURL Enable [<?php if (function_exists('curl_version') !== FALSE) {
+                            }
+                            ?>]<br>
+                            cURL Enable [<?php
+                            if (function_exists('curl_version') !== FALSE) {
                                 echo '<span class="success">PASS</span>';
                                 $curl = 1;
                             } else {
                                 echo '<span class="error">FAIL</span>';
                                 $curl = 0;
-                            } ?>]<br>
-                                PHP GD library Enable [<?php if (extension_loaded('gd') && function_exists('gd_info')) {
+                            }
+                            ?>]<br>
+                            PHP GD library Enable [<?php
+                            if (extension_loaded('gd') && function_exists('gd_info')) {
                                 echo '<span class="success">PASS</span>';
                                 $gd = 1;
                             } else {
                                 echo '<span class="error">FAIL</span>';
                                 $gd = 0;
-                            } ?>]<br>
-                                Config Write Permission [<?php if (is_writable('../config_example.inc.php')) {
+                            }
+                            ?>]<br>
+                            Config Write Permission [<?php
+                            if (is_writable('../config_example.inc.php')) {
                                 echo '<span class="success">PASS</span>';
                                 $config_per = 1;
                             } else {
                                 echo '<span class="error">FAIL</span>';
                                 $config_per = 0;
-                            } ?>]<br>
-                                ci_session Write Permission [<?php if (is_writable('../ci_session/index.html')) {
+                            }
+                            ?>]<br>
+                            ci_session Write Permission [<?php
+                            if (is_writable('../ci_session/index.html')) {
                                 echo '<span class="success">PASS</span>';
                                 $sess_per = 1;
                             } else {
                                 echo '<span class="error">FAIL</span>';
                                 $sess_per = 0;
-                            } ?>]<br>
-                                Cache Write Permission [<?php if (is_writable('../cszcms/cache/index.html')) {
+                            }
+                            ?>]<br>
+                            Cache Write Permission [<?php
+                            if (is_writable('../cszcms/cache/index.html')) {
                                 echo '<span class="success">PASS</span>';
                                 $cache_per = 1;
                             } else {
                                 echo '<span class="error">FAIL</span>';
                                 $cache_per = 0;
-                            } ?>]<br>
-                                DB_Cache Write Permission [<?php if (is_writable('../cszcms/db_cache/index.html')) {
+                            }
+                            ?>]<br>
+                            DB_Cache Write Permission [<?php
+                            if (is_writable('../cszcms/db_cache/index.html')) {
                                 echo '<span class="success">PASS</span>';
                                 $dbcache_per = 1;
                             } else {
                                 echo '<span class="error">FAIL</span>';
                                 $dbcache_per = 0;
-                            } ?>]
+                            }
+                            ?>]
                             <?php
                             if ($sqli == 1 && $php == 1 && $curl == 1 && $gd == 1 && $config_per == 1 && $sess_per == 1 && $cache_per == 1 && $dbcache_per == 1) {
                                 $chk_pass = 1;
@@ -244,7 +256,7 @@ if (!empty($_POST) && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] 
                         <hr>
                     </div>
                 </div>
-                    <?php if ($chk_pass == 1) { ?> 
+                <?php if ($chk_pass == 1) { ?> 
                     <form action="index.php" method="post" accept-charset="utf-8">
                         <div class="row">
                             <div class="col-md-6">
@@ -261,7 +273,7 @@ if (!empty($_POST) && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] 
                                         <input id="dbpass" name="dbpass" type="password" class="form-control" placeholder="Password for DB" required>
                                         <label for="dbname">Database Name*: </label>
                                         <input id="dbname" name="dbname" type="text" class="form-control" placeholder="DB Name for CSZ-CMS" required>
-                                        <br><span class="remark">
+                                        <br><span class="text-info">
                                             <b>Your PHP Version: <u><?php echo phpversion(); ?></u></b><br>
                                             <b>* Required for <u>MySQLi</u> (PHP 5.3.7 or higher, MySQL 5.0 or higher)</b><br>
                                             <b>* Please create the database on your hosting control panel.</b><br><br>
@@ -287,10 +299,10 @@ if (!empty($_POST) && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] 
                                             </span>
                                             <input id="baseurl" name="baseurl" type="text" class="form-control" placeholder="www.ex.com or www.ex.com/subdir" required>
                                         </div><!-- /input-group -->
-                                        <span class="remark"><em>If you want install on sub-directory Example. <b>http://www.ex.com/subdir</b> or <b>http://localhost/subdir</b> on localhost. subdir is directory when you extract file from zip file</em></span><br>
+                                        <span class="text-info"><em>If you want install on sub-directory Example. <b>http://www.ex.com/subdir</b> or <b>http://localhost/subdir</b> on localhost. subdir is directory when you extract file from zip file</em></span><br>
                                         <label for="timezone">Time Zone*: </label>
                                         <input id="timezone" name="timezone" type="text" class="form-control" placeholder="Asia/Bangkok" required>
-                                        <span class="remark"><em>See at <a href="http://php.net/manual/en/timezones.php" target="_blank"><b>Here</b></a> for Timezone list</em></span>
+                                        <span class="text-info"><em>See at <a href="http://php.net/manual/en/timezones.php" target="_blank"><b>Here</b></a> for Timezone list</em></span>
                                     </div>
                                 </div>
                                 <div class="panel panel-default">
@@ -302,30 +314,30 @@ if (!empty($_POST) && $_POST['baseurl'] && $_POST['dbhost'] && $_POST['dbuser'] 
                                         <input id="email" name="email" type="email" class="form-control" placeholder="Email Address for Backend Login">
                                         <label for="password">Password: </label>
                                         <input id="password" name="password" type="password" class="form-control" placeholder="Password for Backend Login">
-                                        <span class="remark"><em>Please <b>blank</b>. If you want default backend login account<br>Email: <b>demo@cszcms.com</b> | Password: <b>123456</b></em></span>
+                                        <span class="text-info"><em>Please <b>blank</b>. If you want default backend login account<br>Email: <b>demo@cszcms.com</b> | Password: <b>123456</b></em></span>
                                     </div>
                                 </div>
                             </div>
                         </div> 
                         <div class="row">
                             <div class="col-md-12 text-right">
-                                <button class="btn btn-lg btn-primary" type="submit" id="submit">Install Now</button>
+                                <input type="submit" class="btn btn-lg btn-primary" id="submitbtn" name="submitbtn" value="Install Now">
                             </div>
                         </div>
                     </form>
-                    <?php }else{ ?>
-                        <div class="row">
-                            <div class="col-md-12 text-center">
-                                <div class="alert alert-danger" role="alert">Your system isn't compatible. Please check your system and reload this page again.</div>
-                            </div>
+                <?php } else { ?>
+                    <div class="row">
+                        <div class="col-md-12 text-center">
+                            <div class="alert alert-danger" role="alert">Your system isn't compatible. Please check your system and reload this page again.</div>
                         </div>
-                    <?php } ?>
+                    </div>
+                <?php } ?>
             <?php } ?>
         </div>
         <!-- End For Content -->
-<?php
-$version = new Version;
-?>
+        <?php
+        $version = new Version;
+        ?>
         <br><br><br>
         <div class="container">
             <footer>
